+++ .formatter.exs
@@ -1,1 +1,1 @@
-[inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]]
+[inputs: ["{mix,.formatter}.exs", "{config,lib,test,benchmark}/**/*.{ex,exs}"]]